Separate struct into a different file
